requirements.txt
flask
flask-sqlalchemy
sqlalchemy
flask-migrate
requests
pytest-flask
jsonschema
